requirements.txt
stable-retro
opencv-python
numpy
keyboard
